refactoring package name stepHandlers to step #9
