+++ src/main/java/sudoku/step/OneChangeStep.java
@@ -1,4 +1,4 @@
-package sudoku.stepHandlers;
+package sudoku.step;
 
 import sudoku.model.Cell;
 import sudoku.model.Sudoku;
