+++ src/main/java/sudoku/strategy/BacktrackLucia.java
@@ -2,8 +2,8 @@ package sudoku.strategy;
 
 import sudoku.model.Cell;
 import sudoku.model.Sudoku;
-import sudoku.stepHandlers.OneChangeStep;
-import sudoku.stepHandlers.Step;
+import sudoku.step.OneChangeStep;
+import sudoku.step.Step;
 
 import java.util.ArrayList;
 import java.util.List;
